Back-end/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
langchain
langchain_community
langfuse